Implemented the rest of the Assembler - Relocation tables, Sections, Symbol tables and the assembler itsself..itselfg,

+++ SSProjekat/Converter.h
@@ -5,8 +5,8 @@
 class Converter {
 public:
 	//returns a string of length exactly 8 (4 bytes)
-	static std::string toHex32(long num);
-	static std::string toHex(long num, int digs);
+	static std::string toHex32(long long num);
+	static std::string toHex(long long num, int digs);
 	static char hexChar(int dig);
 	static std::string toLower(std::string s);
 	static std::string removeBlanks(std::string s);
